Added Album creation, mega_run, changed lang codes

+++ routers/album.py
@@ -6,7 +6,8 @@ from aiogram.fsm.context import FSMContext
 from aiogram.fsm.state import State, StatesGroup
 import json
 
-from create_bot import logger, bot
+from create_bot import logger, bot, cm
+from creation.CreationManager import Album
 from messages import album as messages
 from messages import core as messages_core
 from keyboards import album as keyboards
@@ -15,6 +16,7 @@ from utility.template_images import get_image
 
 from logic.core import *
 from logic.album import *
+from creation.asyncio import get_unique_id
 
 
 router = Router(name='album')
@@ -37,6 +39,10 @@ async def query_create_album(query: CallbackQuery, state: FSMContext):
 
     if not await check_sub_or_free_albums(user):
         await query.answer(messages.no_free_albums(lang), show_alert=True)
+        return
+
+    if cm.in_album_queue(user.id):
+        await query.answer(messages.album_query_block(lang), show_alert=True)
         return
 
     images = get_image('templates_album')
@@ -80,7 +86,7 @@ async def query_wait_for_template(query: CallbackQuery, state: FSMContext):
     
     last_message = await query.message.edit_text(
         text=text+messages.wait_for_photo(lang),
-        reply_markup=None
+        reply_markup=keyboards_core.go_back(lang)
     )
 
     data['last_message_id'] = last_message.message_id
@@ -92,18 +98,25 @@ async def query_wait_for_template(query: CallbackQuery, state: FSMContext):
     logger.info(f'@{user.username} chose album template {tmp}')
 
 
-@router.message(StateFilter(CreationStates.wait_for_first_photo), F.photo)
+@router.message(StateFilter(CreationStates.wait_for_first_photo), F.document)
 async def message_wait_for_first_photo(message: Message, state: FSMContext):
     user = message.from_user
     await update_user(user)
     lang = await get_language(user)
     data = await state.get_data()
 
-    file_id = message.photo[-1].file_id
+    if not (message.document and 'image' in message.document.mime_type):
+        await message.answer(
+            text=messages.wrong_photo_format(lang)
+        )       
+        return
+
+    file_id = message.document.file_id
     data['photos'] = [file_id]
 
     last_message = await message.answer(
-        text=messages.wait_second_photo(lang)
+        text=messages.wait_second_photo(lang),
+        reply_markup=keyboards_core.go_back(lang)
     )
     
     await bot.delete_message(user.id, data['last_message_id'])
@@ -115,21 +128,26 @@ async def message_wait_for_first_photo(message: Message, state: FSMContext):
     logger.info(f'@{user.username} sent first photo, id = {file_id}')
 
 
-@router.message(StateFilter(CreationStates.wait_for_single_photo, CreationStates.wait_for_second_photo), F.photo)
+@router.message(StateFilter(CreationStates.wait_for_single_photo, CreationStates.wait_for_second_photo), F.document)
 async def message_wait_for_singe_or_second_photo(message: Message, state: FSMContext):
     user = message.from_user
     await update_user(user)
     lang = await get_language(user)
     data = await state.get_data()
 
-    file_id = message.photo[-1].file_id
+    if not (message.document and 'image' in message.document.mime_type):
+        await message.answer(
+            text=messages.wrong_photo_format(lang)
+        )       
+        return
+
+    file_id = message.document.file_id
     if 'photos' in data.keys():
         data['photos'].append(file_id)
     else:
         data['photos'] = [file_id]
 
     last_message = await message.answer(
-        # photo=tmp_photo...                                                      <--------------------------------------- TODO
         text=messages.create_album_approve(lang, data),
         reply_markup=keyboards.create_album_approve(lang)
     )
@@ -170,21 +188,42 @@ async def query_wait_for_approve(query: CallbackQuery, state: FSMContext):
             await query.answer(messages.no_free_albums(lang), show_alert=True)
             return
 
+    if cm.in_album_queue(user.id):
+        await query.answer(messages.album_query_block(lang), show_alert=True)
+        return
+
     await query.message.edit_text(
         text=messages.creation_end(lang, 20, 0),
         reply_markup=keyboards_core.go_back(lang)
     )
 
-    await query.message.answer(
-        text='Создание альбомов пока не готово 🛠️'
-    )
+    photos = data['photos']
 
-    # await query.message.answer(
-    #     text=(
-    #         f'```json\n{json.dumps(data, indent=4, ensure_ascii=False)}\n```'
-    #     ),
-    #     parse_mode='MarkdownV2'
-    # )
+    files = [f'creation/img/{user.id}_{x}.jpeg' for x in photos]
 
-    await state.clear()
+    print('Album creation started')
+
+    await bot.download(photos[0], files[0])
+    if len(photos) > 1:
+        await bot.download(photos[1], files[1])
+    
+    if len(files) == 1: files.append(None) 
+    try:
+        await use_free_albums(user)
+        unique_id = get_unique_id()
+        alb_file = await cm.createAlbum(Album(user.id, unique_id, data['template'], files[0], files[1]))
+        
+        await query.message.answer_photo(
+            photo=BufferedInputFile(file=open(alb_file, 'rb').read(), filename='album for you'),
+            caption=messages.album_ready(lang)
+        )
+
+    except Exception as e:
+        print(e)
+        await add_free_albums(user)
+        if 'VOICE_MESSAGES_FORBIDDEN' in e.__str__():
+            await query.message.answer(messages_core.voice_forbidden(lang))
+        else:
+            await query.message.answer(messages_core.error(lang))
+
     logger.info(f'@{user.username} started album creation')
